project/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
httpx==0.25.0
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-asyncio==0.21.1
SQLAlchemy==2.0.22
numpy==1.26.0

